Day idfk - Fine Dining Page.
I can load the navigation bar in, but it messes with the css. Have to revise that.

Photo ring is less of a ring and more of an arc display. Need to figure out how I should go about fixin it.

Slideshow on the bottom is a fuckin mess, fix it.